Fixed warmup bug
For high probabilities of P_h, the neuron almost always looks when considered after exactly x time steps. To overcome this, the warmup value will be uniformly distributed 25% around the specified value. This is to simulate the injection of a current at some arbitrary time where conditions are more or less in a steady state